# Attach last error to the state