This fixes https://github.com/aarzilli/golua/issues/24 by adding a `lastErr` attribute to the state, we can track the error that was received by the error handler while still allowing the error handler function to return. Since the function returns, the stack is properly cleaned up and both the `"C stack overflow"` and `"error in error handling"` errors go away.

This fix also improves performance:

Before:

```
BenchmarkFix-8          10000        131910 ns/op       45529 B/op         897 allocs/op
PASS
ok      github.com/andrewhare/tmp    1.340s
```

After:

```
BenchmarkFix-8         300000          4258 ns/op         552 B/op          14 allocs/op
PASS
ok      github.com/andrewhare/tmp    1.348s
```

Running the following benchmark:

```go
package main

import (
    "testing"

    "github.com/andrewhare/golua/lua"
)

var script = `
function doit()
    x[1] = 2
end
`

func BenchmarkFix(b *testing.B) {
    L := lua.NewState()
    L.OpenLibs()
    if err := L.DoString(script); err != nil {
        panic(err)
    }
    for n := 0; n < b.N; n++ {
        top := L.GetTop()
        L.GetGlobal("doit")
        err := L.Call(0, lua.LUA_MULTRET)
        if err == nil {
            panic("should have error")
        }
        L.SetTop(top)
    }
}
```