@aarzilli 2019-06-07
Lua errors can't be allowed to complete, otherwise they'll end up executing a longjump, which will skip over a bunch of go frames, which isn't allowed by go runtime. See CI failure.