@andrewhare 2019-06-17
@aarzilli I've pushed an experimental approach, basically wrapping each API call in `pcall`. While this solves the segfault, it still doesn't bubble the wrapped call error back up the way the panic did before so this isn't complete.

I'm probably not going to put any more work into this than I have since my original fix works for my use-case, so feel free to do whatever you want with this PR.